+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Extract click methods to `ClickMethods` module
 * Increase test coverage of `ClickMethods` module
 * Remove `WebDriver#click_and_wait`
+* `WebDriver#click_on_locator` raise `Selenium::WebDriver::Error::ElementNotVisibleError`
+  instead of `RuntimeError`
 
 ## 0.3.2 (2020-06-30)