+++ CHANGELOG.md
@@ -14,7 +14,8 @@
 * `WebDriver#click_on_locator` raise `Selenium::WebDriver::Error::ElementNotVisibleError`
   instead of `RuntimeError`
 * `WebDriver#click_on_displayed` raise a correct exception type if failed
-* Remove unused `WebDrvier#click_on_one_of_several_by_text`
+* Remove unused `WebDrvier#click_on_one_of_several_by_text`,
+  `WebDriver#click_on_one_of_several_xpath_by_number`
 
 ## 0.3.2 (2020-06-30)