+++ CHANGELOG.md
@@ -15,7 +15,8 @@
   instead of `RuntimeError`
 * `WebDriver#click_on_displayed` raise a correct exception type if failed
 * Remove unused `WebDrvier#click_on_one_of_several_by_text`,
-  `WebDriver#click_on_one_of_several_xpath_by_number`
+  `WebDriver#click_on_one_of_several_xpath_by_number`, 
+  `WebDriver#left_mouse_click`
 
 ## 0.3.2 (2020-06-30)