+++ CHANGELOG.md
@@ -17,6 +17,9 @@
 * Remove unused `WebDrvier#click_on_one_of_several_by_text`,
   `WebDriver#click_on_one_of_several_xpath_by_number`, 
   `WebDriver#left_mouse_click`
+* `WebDriver#right_click` wait to element to appear
+* Remove `WebDriver#context_click_on_locator` -
+  use `WebDriver#right_click` instead
 
 ## 0.3.2 (2020-06-30)